@@ -12,6 +12,7 @@
   },
   "scripts": {
     "build": "ember build --environment=production",
+    "watch": "ember build --watch",
     "lint": "npm-run-all --aggregate-output --continue-on-error --parallel 'lint:!(fix)'",
     "lint:fix": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*:fix",
     "lint:hbs": "ember-template-lint .",
@@ -28,6 +29,12 @@
     "@ember/test-helpers": "^2.1.4",
     "@glimmer/component": "^1.0.3",
     "@glimmer/tracking": "^1.0.3",
+    "@types/ember": "^3.16.2",
+    "@types/ember-data": "^3.16.8",
+    "@types/ember-qunit": "^3.4.13",
+    "@types/ember-resolver": "^5.0.10",
+    "@types/ember__test-helpers": "^1.7.3",
+    "@types/qunit": "^2.11.1",
     "babel-eslint": "^10.1.0",
     "broccoli-asset-rev": "^3.0.0",
     "ember-auto-import": "^1.10.1",
@@ -39,6 +46,9 @@
     "ember-cli-inject-live-reload": "^2.0.2",
     "ember-cli-sri": "^2.1.1",
     "ember-cli-terser": "^4.0.1",
+    "ember-cli-typescript": "^4.0.0",
+    "ember-cli-typescript-blueprints": "^3.0.0",
+    "ember-css-modules": "^1.3.3",
     "ember-data": "~3.24.0",
     "ember-electron": "3.0.0-beta.4",
     "ember-export-application-global": "^2.0.1",
@@ -50,7 +60,6 @@
     "ember-resolver": "^8.0.2",
     "ember-source": "~3.24.0",
     "ember-template-lint": "^2.15.0",
-    "ember-welcome-page": "^4.0.0",
     "eslint": "^7.17.0",
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-ember": "^10.1.1",
@@ -58,14 +67,21 @@
     "eslint-plugin-prettier": "^3.3.1",
     "loader.js": "^4.7.0",
     "npm-run-all": "^4.1.5",
+    "postcss-import": "^12.0.1",
+    "postcss-nested": "^4.2.1",
     "prettier": "^2.2.1",
     "qunit": "^2.13.0",
-    "qunit-dom": "^1.6.0"
+    "qunit-dom": "^1.6.0",
+    "tailwindcss": "2.0.0-compat",
+    "typescript": "^4.1.3"
   },
   "engines": {
-    "node": "10.* || >= 12"
+    "node": ">= 12"
   },
   "ember": {
     "edition": "octane"
+  },
+  "volta": {
+    "node": "12.20.1"
   }
 }
