@@ -1,20 +1,17 @@
 /* eslint-disable no-console */
-const {
-  default: installExtension,
-  EMBER_INSPECTOR,
-} = require('electron-devtools-installer');
-const { pathToFileURL } = require('url');
-const { app, BrowserWindow } = require('electron');
-const path = require('path');
-const isDev = require('electron-is-dev');
-const handleFileUrls = require('./handle-file-urls');
+// import installExtension, { EMBER_INSPECTOR } from 'electron-devtools-installer';
+import { pathToFileURL } from 'url';
+import { app, BrowserWindow } from 'electron';
+import * as path from 'path';
+import * as isDev from 'electron-is-dev';
+import handleFileUrls from './handle-file-urls';
 
 const emberAppDir = path.resolve(__dirname, '..', 'ember-dist');
 const emberAppURL = pathToFileURL(
   path.join(emberAppDir, 'index.html')
 ).toString();
 
-let mainWindow = null;
+let mainWindow: BrowserWindow | null = null;
 
 // Uncomment the lines below to enable Electron's crash reporter
 // For more information, see http://electron.atom.io/docs/api/crash-reporter/
@@ -32,26 +29,36 @@ app.on('window-all-closed', () => {
 });
 
 app.on('ready', async () => {
-  if (isDev) {
-    try {
-      require('devtron').install();
-    } catch (err) {
-      console.log('Failed to install Devtron: ', err);
-    }
-    try {
-      await installExtension(EMBER_INSPECTOR);
-    } catch (err) {
-      console.log('Failed to install Ember Inspector: ', err);
-    }
-  }
+  // Mostly broken: Because of some changes in Chrome, in recent Electron versions (>=9 I think) they can't be loaded with file: protocol https://github.com/electron/electron/issues/24011),
+  // and also devtron doesn't work for other reasons (https://github.com/electron/electron/issues/23676)
+  // and also devtron has been mostly abandoned (https://github.com/electron-userland/devtron/issues/200).
+  //
+  // if (isDev) {
+  //   try {
+  //     require('devtron').install();
+  //   } catch (err) {
+  //     console.log('Failed to install Devtron: ', err);
+  //   }
+  //   try {
+  //     await installExtension(EMBER_INSPECTOR);
+  //   } catch (err) {
+  //     console.log('Failed to install Ember Inspector: ', err);
+  //   }
+  // }
 
   await handleFileUrls(emberAppDir);
 
   mainWindow = new BrowserWindow({
     width: 800,
     height: 600,
+    webPreferences: {
+      contextIsolation: true,
+    },
   });
-  mainWindow.webContents.openDevTools();
+
+  if (isDev) {
+    mainWindow.webContents.openDevTools();
+  }
 
   // If you want to open up dev tools programmatically, call
   // mainWindow.openDevTools();
